step_1_rl/requirements.txt
absl-py @ file:///C:/b/abs_14wchyjaia/croot/absl-py_1714140489450/work
ale-py==0.10.1
alembic==1.14.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
bottle==0.13.2
box2d-py==2.3.5
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
category-encoders==2.6.3
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
chex==0.1.88
cloudpickle==3.1.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
colorlog==6.9.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
cycler==0.12.1
Cython==0.29.37
dash==2.18.1
debugpy @ file:///C:/b/abs_bf9oo2vhxp/croot/debugpy_1736269476451/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1733236420667/work
etils==1.11.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1733208806608/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1733569351617/work
Farama-Notifications==0.0.4
fasteners==0.19
filelock @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/filelock_1701807523603/work
Flask==3.0.3
flax==0.10.2
fonttools==4.55.3
fsspec==2024.12.0
glfw==2.8.0
greenlet==3.1.1
grpcio==1.70.0
gymnasium==1.0.0
humanize==4.11.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1734788189411/work
jax==0.4.38
jaxlib==0.4.38
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2 @ file:///C:/b/abs_ddyeze2h6_/croot/jinja2_1730903330414/work
joblib==1.4.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///C:/b/abs_beftpbuevw/croot/jupyter_core_1718818307097/work
keyboard==0.13.5
kiwisolver==1.4.8
lightgbm==4.5.0
Mako==1.3.8
Markdown @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markdown_1699503092856/work
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markupsafe_1707425732791/work
matplotlib==3.10.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.0
moviepy==2.1.1
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
msgpack==1.1.0
mujoco==3.2.6
mujoco-py==2.1.2.14
natsort==8.4.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optuna==4.2.0
optuna-dashboard==0.17.0
orbax-checkpoint==0.11.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1733232627818/work
plotly-resampler==0.10.0
pmdarima==2.0.4
proglog==0.1.10
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1737453357274/work
protobuf==4.25.3
psutil @ file:///C:/b/abs_b5gv3mn55h/croot/psutil_1736371546320/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pycparser==2.22
pygame==2.6.1
Pygments==2.18.0
PyOpenGL==3.1.7
pyparsing==3.2.1
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq @ file:///C:/b/abs_f3yte6j5yn/croot/pyzmq_1734711069724/work
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.0
setuptools==72.1.0
simplejson==3.19.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
SQLAlchemy==2.0.37
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
swig==4.3.0
sympy==1.13.1
tbats==1.1.3
tensorboard @ file:///C:/b/abs_8bayrkofnj/croot/tensorboard_1723042051994/work/tensorboard-2.17.0-py3-none-any.whl#sha256=859a499a9b1fb68a058858964486627100b71fcb21646861c61d31846a6478fb
tensorboard_data_server @ file:///C:/b/abs_96oxnengxy/croot/tensorboard-data-server_1724172563443/work/tensorboard_data_server-0.7.0-py3-none-any.whl#sha256=0d8e38f3cc912719d7e458526106e6d9b2f0a1216a702f53d3f8da5db3cd1f56
tensorstore==0.1.71
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///C:/b/abs_7cyu943ybx/croot/tornado_1733960510898/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2024.2
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug @ file:///C:/b/abs_c7bupijx2_/croot/werkzeug_1737448709012/work
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
